requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Werkzeug==3.0.1
python-dotenv==1.0.0
# Core dependencies
pymupdf==1.24.14
pymupdf4llm==0.0.17
openai
nltk
lexicalrichness
sqlalchemy
requests
asyncio
aiohttp
